+++ packages/registry/components/examples/papi/example-tx-button.tsx
@@ -7,7 +7,7 @@ import { useChainIds } from "@reactive-dot/react";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import { TxButton } from "@/registry/polkadot-ui/blocks/tx-button/components/tx-button.papi";
 import { Binary } from "polkadot-api";
-import { config } from "@/registry/polkadot-ui/reactive-dot.config";
+import { config } from "@/registry/polkadot-ui/lib/reactive-dot.config";
 import { ChainId } from "@reactive-dot/core";
 import { usePapi } from "@/registry/polkadot-ui/lib/polkadot-provider.papi";
 import { Suspense } from "react";
