+++ packages/registry/registry/polkadot-ui/blocks/balance-display/balance-display.papi.tsx
@@ -10,7 +10,7 @@ import {
   useNativeBalance,
 } from "@/registry/polkadot-ui/hooks/use-asset-balance.papi";
 import { NATIVE_TOKEN_KEY } from "@/registry/polkadot-ui/lib/utils.dot-ui";
-import { ChainIdsWithPalletAssets } from "@/registry/polkadot-ui/reactive-dot.config";
+import { ChainIdsWithPalletAssets } from "@/registry/polkadot-ui/lib/reactive-dot.config";
 
 export type BalanceDisplayProps = Omit<
   BalanceDisplayBaseProps,
