+++ packages/registry/registry/polkadot-ui/blocks/select-token/select-token-dialog.papi.tsx
@@ -27,7 +27,7 @@ import {
   NATIVE_TOKEN_ID,
 } from "@/registry/polkadot-ui/lib/utils.dot-ui";
 import { ClientConnectionStatus } from "@/registry/polkadot-ui/lib/types.dot-ui";
-import { config } from "@/registry/polkadot-ui/reactive-dot.config";
+import { config } from "@/registry/polkadot-ui/lib/reactive-dot.config";
 import type { ChainId } from "@reactive-dot/core";
 
 export type SelectTokenDialogProps = Omit<
