+++ packages/registry/registry/polkadot-ui/lib/utils.dot-ui.ts
@@ -1,6 +1,4 @@
-import { decodeAddress, encodeAddress } from "@polkadot/keyring";
-import { ethers } from "ethers";
-import { TokenInfo } from "@/registry/polkadot-ui/lib/types.dot-ui";
+import type { TokenInfo } from "@/registry/polkadot-ui/lib/types.dot-ui";
 
 // the key for the native token in the balances object, if -1 is used somewhere as assetId, the
 export const NATIVE_TOKEN_KEY = -1;
@@ -62,69 +60,6 @@ export interface ValidationResult {
   type: "ss58" | "eth" | "unknown";
   error?: string;
   normalizedAddress?: string;
-}
-
-export function validateAddress(
-  address: string,
-  format: "eth" | "ss58" | "both",
-  ss58Prefix: number = 42
-): ValidationResult {
-  if (!address.trim()) {
-    return { isValid: false, type: "unknown", error: "Address is required" };
-  }
-
-  // SS58 validation
-  if (format === "ss58" || format === "both") {
-    try {
-      const decoded = decodeAddress(address);
-
-      // Check if the decoded address has the proper length (32 bytes for a public key)
-      if (decoded.length !== 32) {
-        throw new Error("Invalid address length");
-      }
-
-      const encoded = encodeAddress(decoded, ss58Prefix);
-      return {
-        isValid: true,
-        type: "ss58",
-        normalizedAddress: encoded,
-      };
-    } catch {
-      if (format === "ss58") {
-        return {
-          isValid: false,
-          type: "unknown",
-          error: "Invalid Polkadot address format",
-        };
-      }
-    }
-  }
-
-  // Ethereum validation
-  if (format === "eth" || format === "both") {
-    try {
-      const normalized = ethers.getAddress(address);
-      return {
-        isValid: true,
-        type: "eth",
-        normalizedAddress: normalized,
-      };
-    } catch {
-      if (format === "eth") {
-        return {
-          isValid: false,
-          type: "unknown",
-          error: "Invalid Ethereum address format",
-        };
-      }
-    }
-  }
-
-  return {
-    isValid: false,
-    type: "unknown",
-    error: "Invalid address format",
-  };
 }
 
 export function truncateAddress(
